seppareted mappedlist into file
